Set VS Command in compiler.py
This forked version of PySimlink is designed to fix a bug which causes the compiler not to work, owing to the inability to find the compile .exe file (for Windows, nmake) as it is not on the path.

In this fix, a line is added to the method _build of the compiler class that calls VsDemCmd.bat, a batch file that sets the environment within a command prompt such that it mirrors Visual Studio, enabling key VS compiler files to be found.